requirements-dev.txt
pytest
pytest-cov
pyhamcrest
doublex
coverage
ruff
mypy
